:construction: Game, Ball, Paddle 타입 수정

+++ back/pong_game/module/Paddle.py
@@ -3,11 +3,11 @@ from .Game import FIELD_LENGTH
 
 class Paddle:
     def __init__(self):
-        self.position_x: int = 0
-        self.position_y: int = 0
-        self.position_z: int = FIELD_LENGTH / 2
-        self.width: int = 0
-        self.speed: int = 0
+        self.position_x: float = 0
+        self.position_y: float = 0
+        self.position_z: float = FIELD_LENGTH / 2
+        self.width: float = 0
+        self.speed: float = 0
 
     def reset_position(self):
         self.position_x = 0
